requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0gunicorn==23.0.0openai>=1.8.3
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
python-dotenv==1.0.0
pydantic>=2.0.0
numpy>=2.1.0
torch>=2.6.0
transformers>=4.40.0
PyPDF2>=3.0.1
# Optional legacy / specialized packages retained from project
llama-index>=0.10.44
jsonargparse>=4.29.0
streamlit>=1.28.1
requests>=2.32.4
beautifulsoup4>=4.12.2
urllib3>=2.0.7